functions/ocr_function/requirements.txt
zcatalyst-sdk==0.0.2
XlsxWriter==3.2.0
numpy==1.26.4
pandas==2.0.3
Flask-Cors==4.0.1
PyPDF2==3.0.1